Edit sample files
Added in instructions to run manifests through JSONlint. It's an
excellent tool, and even cleans up whitespace and simple formatting
errors, such as putting brackets on their own line when it's not
needed. With any luck, it'll prevent future pull requests from breaking
manifests.js, as well as a developer's own manifest file.